pyproject.toml
[project]
name = "noip-automation"
version = "0.1.0"
description = ""
authors = [
    {name = "sethorpe",email = "oluseyi.thorpe@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright (>=1.57.0,<2.0.0)",
    "pyotp (>=2.9.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
